Added a pause function.
I added a bool which when set to true, will pause the game and stop the player from moving. Currently contributed on the Timer UI aspect, edited the testing level a bit. and kept things more organized.

I made some edit to the player movement script to make it "pauseable"